arreglando lo del borrado de logs

+++ TiX/PythonApp/ServerApp/udpServerTiempos.py
@@ -26,6 +26,7 @@ TEST_SERVER_HOST = config.get("TiXServer", "TEST_SERVER_HOST") #TODO: Change TES
 TEST_SERVER_PORT = config.getint("TiXServer", "TEST_SERVER_PORT")
 installDirUnix = config.get("TiXServer", "installDirUnix")
 tixBaseUrl = config.get("TiXServer", "tixBaseUrl")
+modo_debug = config.get("TixServer", "modo_debug")
 
 sys.path.append('/home/pfitba/ServerAppProduction/data_processing/')
 import completo_III
@@ -86,7 +87,7 @@ def remove_old_files(dirpath, client_msg_filename):
             curr_file_datetime = datetime.datetime.fromtimestamp(float(file_name.split("_")[1]))
             lapsed_time = (log_datetime-curr_file_datetime).total_seconds()
             try:
-                if abs(lapsed_time) > 7200: #1 hora y 30 minutos de ventana
+                if abs(lapsed_time) > 7200: #2 horas
                     os.remove(dirpath + "/" + file_name)
                     logger.info("Eliminando log antiguo: " + dirpath + "/" + file_name)
 
@@ -106,7 +107,8 @@ def remove_1h_files(dirpath):
         logger.debug('Hay ' + str(len(a)) + ' arhivos en el directorio ' + dirpath)
         file_to_remove = a.pop(0)
         logger.info("Eliminando log antiguo: " + dirpath + "/" + file_to_remove)
-        os.remove(dirpath + "/" + file_to_remove)
+        if modo_debug == True and modo_debug == True
+            os.remove(dirpath + "/" + file_to_remove)
         a.sort(key=lambda x: os.stat(os.path.join(dirpath, x)).st_mtime)
 
 class ThreadingUDPRequestHandler(SocketServer.BaseRequestHandler):
@@ -212,7 +214,7 @@ class ThreadingUDPRequestHandler(SocketServer.BaseRequestHandler):
 
                             # Remove 10 oldest logs
                             for count in range(0,9):
-                                if os.path.isfile(files_to_process[count]) == True:
+                                if os.path.isfile(files_to_process[count]) == True and modo_debug == True:
                                     os.remove(files_to_process[count])
                             try:
                                 new_isp_name = info.pais_num_name_nic(client_ip, 'EN' )[1]
